Moved protocol decoder so separate file
